+++ app/src/main/java/com/yapp/breake/di/UseCaseModule.kt
@@ -2,24 +2,24 @@ package com.yapp.breake.di
 
 import com.yapp.breake.domain.usecase.CreateNewGroupUseCase
 import com.yapp.breake.domain.usecase.DecideNextDestinationFromPermissionUseCase
-import com.yapp.breake.domain.usecase.DecideNextDestinationFromPermissionUseCaseImpl
+import com.yapp.breake.domain.usecaseImpl.DecideNextDestinationFromPermissionUseCaseImpl
 import com.yapp.breake.domain.usecase.DecideStartDestinationUseCase
-import com.yapp.breake.domain.usecase.DecideStartDestinationUseCaseImpl
+import com.yapp.breake.domain.usecaseImpl.DecideStartDestinationUseCaseImpl
 import com.yapp.breake.domain.usecase.DeleteAccountUseCase
-import com.yapp.breake.domain.usecase.DeleteAccountUseCaseImpl
+import com.yapp.breake.domain.usecaseImpl.DeleteAccountUseCaseImpl
 import com.yapp.breake.domain.usecase.DeleteGroupUseCase
 import com.yapp.breake.domain.usecase.FindAppGroupUseCase
 import com.yapp.breake.domain.usecase.GetNicknameUseCase
 import com.yapp.breake.domain.usecase.LoginUseCase
 import com.yapp.breake.domain.usecase.LogoutUseCase
-import com.yapp.breake.domain.usecase.LogoutUseCaseImpl
+import com.yapp.breake.domain.usecaseImpl.LogoutUseCaseImpl
 import com.yapp.breake.domain.usecase.ResetAppGroupUsecase
-import com.yapp.breake.domain.usecase.ResetAppGroupUsecaseImpl
+import com.yapp.breake.domain.usecaseImpl.ResetAppGroupUsecaseImpl
 import com.yapp.breake.domain.usecase.SetAlarmUseCase
 import com.yapp.breake.domain.usecase.SetBlockingAlarmUseCase
 import com.yapp.breake.domain.usecase.SetSnoozeAlarmUseCase
 import com.yapp.breake.domain.usecase.StoreOnboardingCompletionUseCase
-import com.yapp.breake.domain.usecase.StoreOnboardingCompletionUseCaseImpl
+import com.yapp.breake.domain.usecaseImpl.StoreOnboardingCompletionUseCaseImpl
 import com.yapp.breake.domain.usecase.UpdateNicknameUseCase
 import com.yapp.breake.domain.usecaseImpl.CreateNewGroupUseCaseImpl
 import com.yapp.breake.domain.usecaseImpl.DeleteGroupUseCaseImpl
