+++ domain/src/main/java/com/yapp/breake/domain/usecaseImpl/DecideNextDestinationFromPermissionUseCaseImpl.kt
@@ -1,7 +1,8 @@
-package com.yapp.breake.domain.usecase
+package com.yapp.breake.domain.usecaseImpl
 
 import com.yapp.breake.core.model.user.Destination
 import com.yapp.breake.domain.repository.SessionRepository
+import com.yapp.breake.domain.usecase.DecideNextDestinationFromPermissionUseCase
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.runBlocking
 import javax.inject.Inject
