+++ domain/src/main/java/com/yapp/breake/domain/usecaseImpl/LogoutUseCaseImpl.kt
@@ -1,32 +1,28 @@
-package com.yapp.breake.domain.usecase
+package com.yapp.breake.domain.usecaseImpl
 
 import com.yapp.breake.core.model.user.Destination
 import com.yapp.breake.domain.repository.SessionRepository
+import com.yapp.breake.domain.usecase.LogoutUseCase
 import javax.inject.Inject
 
-class DeleteAccountUseCaseImpl @Inject constructor(
+class LogoutUseCaseImpl @Inject constructor(
 	private val sessionRepository: SessionRepository,
-) : DeleteAccountUseCase {
+) : LogoutUseCase {
 	override suspend fun invoke(onError: suspend (Throwable) -> Unit): Destination {
 		return try {
-			sessionRepository.clearRemoteAccount(
+			sessionRepository.clearEntireDataStore(
 				onError = { throwable ->
 					onError(throwable)
-					throw ServerException()
+					throw LocalException()
 				},
 			)
-			sessionRepository.clearEntireDataStore(onError = { throwable ->
-				onError(throwable)
-				throw LocalException()
-			})
 			Destination.Login
-		} catch (_: Exception) {
-			Destination.NotChanged
+		} catch (_: LocalException) {
+			Destination.PermissionOrHome
 		}
 	}
 
 	companion object {
-		class ServerException : Exception()
 		class LocalException : Exception()
 	}
 }
